SERVER-78675 move getMaxNumberOfTransactionOperationsInSingleOplogEntry() and getMaxSizeOfTransactionOperationsInSingleOplogEntryBytes() from OpObserverImpl to TransactionParticipant
