+++ src/mongo/db/transaction/transaction_participant.h
@@ -1347,4 +1347,27 @@ private:
     bool _hasSeenIncomingConflictingRetryableTransaction{false};
 };
 
+/**
+ * Returns maximum number of operations to pack into a single oplog entry,
+ * when multi-oplog format for transactions is in use.
+ *
+ * Stop packing when either number of transaction operations is reached, or when the
+ * next one would make the total size of operations larger than the maximum BSON Object
+ * User Size. We rely on the headroom between BSONObjMaxUserSize and
+ * BSONObjMaxInternalSize to cover the BSON overhead and the other "applyOps" entry
+ * fields. But if a single operation in the set exceeds BSONObjMaxUserSize, we still fit
+ * it, as a single max-length operation should be able to be packed into an "applyOps"
+ * entry.
+ */
+std::size_t getMaxNumberOfTransactionOperationsInSingleOplogEntry();
+
+/**
+ * Returns maximum size (bytes) of operations to pack into a single oplog entry,
+ * when multi-oplog format for transactions is in use.
+ *
+ * Refer to getMaxNumberOfTransactionOperationsInSingleOplogEntry() comments for a
+ * description on packing transaction operations into "applyOps" entries.
+ */
+std::size_t getMaxSizeOfTransactionOperationsInSingleOplogEntryBytes();
+
 }  // namespace mongo
